% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundary_val}
\alias{boundary_val}
\title{Boundary values files}
\format{
A spatial data frame with 1 rows and 4 columns:
\describe{
  \item{id}{ID of boundary}
  \item{lat}{Latitude}
  \item{long}{Longitude}
  ...
}
}
\source{
\url{https://mohp.gov.np}
}
\usage{
boundary_val
}
\description{
A dataset with value of shape files with boundary values
}
\keyword{datasets}
