% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{district_df}
\alias{district_df}
\title{District shape files}
\format{
A spatial data frame with 77 rows and 2 columns:
\describe{
  \item{id}{ID Order by Province-wise District}
  \item{district_id}{Id of each District}
  \item{geometry}{Geometry coordinates of Districts}
  ...
}
}
\source{
\url{https://mohp.gov.np}
}
\usage{
district_df
}
\description{
A dataset map shape files with values of district
}
\keyword{datasets}
