% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{district_map}
\alias{district_map}
\title{Create a District Map}
\usage{
district_map(
  main_df = nepalmap::district_df,
  data = nepalmap::district_val,
  fill = "district",
  label = "district",
  color = "white",
  x = "lat",
  y = "long",
  size = 1,
  legend = ""
)
}
\arguments{
\item{main_df}{The main data frame for creating the map border of district}

\item{data}{The data used for value inputs}

\item{fill}{The color fill shown inside the map}

\item{label}{The label for the province}

\item{color}{The border color}

\item{x}{The latitude by default}

\item{y}{The longitude by default}

\item{size}{Size of label text, 1 by default}

\item{legend}{Legend is empty by default}
}
\value{
a ggplot2 object
}
\description{
Create a District Map
}
\examples{

\dontrun{

district_map()

}

}
