% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{province_df}
\alias{province_df}
\title{Province shape files}
\format{
A spatial data frame with 7 rows and 4 columns:
\describe{
  \item{id}{province number in digit}
  \item{name}{province name}
  \item{capital}{province capital}
  \item{geometry}{Geometry coordinates of Provinces}
  ...
}
}
\source{
\url{https://mohp.gov.np}
}
\usage{
province_df
}
\description{
A dataset map shape files with values of province
}
\keyword{datasets}
