% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{province_map}
\alias{province_map}
\title{Create a province map}
\usage{
province_map(
  main_df = nepalmap::province_df,
  data = nepalmap::province_val,
  fill = "id",
  label = "name",
  color = "white",
  x = "lat",
  y = "long",
  size = 1,
  legend = ""
)
}
\arguments{
\item{main_df}{The main data frame for creating the map border of province}

\item{data}{The data used for value inputs}

\item{fill}{The color fill shown inside the map}

\item{label}{The label for the province}

\item{color}{The border color}

\item{x}{The latitude by default}

\item{y}{The longitude by default}

\item{size}{Size of label text, 1 by default}

\item{legend}{Legend is empty by default}
}
\value{
a ggplot2 object
}
\description{
Create a province map
}
\examples{
 

\dontrun{

province_map()

}
}
