% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{province_val}
\alias{province_val}
\title{Province values files}
\format{
A spatial data frame with 7 rows and 15 variables:
\describe{
  \item{id}{ID of province}
  \item{name}{Name of Province}
  \item{capital}{Capital of Province}
  \item{lat}{Latitude}
  \item{long}{Longitude}
  ...
}
}
\source{
\url{https://mohp.gov.np}
}
\usage{
province_val
}
\description{
A dataset with value of shape files with values of province
}
\keyword{datasets}
