% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundary_df}
\alias{boundary_df}
\title{Boundary shape files}
\format{
A spatial data frame with 1 rows and 4 columns:
\describe{
  \item{id}{Id value}
  \item{geometry}{Geometry coordinates of Boundary}
  ...
}
}
\source{
\url{https://mohp.gov.np}
}
\usage{
boundary_df
}
\description{
A dataset map shape files with values of international boundary of Nepal
}
\keyword{datasets}
