% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundary_val}
\alias{boundary_val}
\title{Boundary values files}
\format{
A spatial data frame with 1 rows and 15 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}
}
\source{
\url{https://mohp.gov.np}
}
\usage{
boundary_val
}
\description{
A dataset with value of shape files with values of boundary
}
\keyword{datasets}
