% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{district_df}
\alias{district_df}
\title{District shape files}
\format{
A spatial data frame with 77 rows and 15 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}
}
\source{
\url{https://mohp.gov.np}
}
\usage{
district_df
}
\description{
A dataset map shape files with values of district
}
\keyword{datasets}
