% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{district_val}
\alias{district_val}
\title{District values files}
\format{
A spatial data frame with 77 rows and 15 variables:
\describe{
  \item{id}{ID order by Province-wise Districts}
  \item{district_id}{ID of each Districts}
  \item{province_n}{Province number in digit}
  \item{province}{Province name}
  \item{district}{District name}
  \item{district_n}{Distric name in Nepali}
  \item{lat}{Latitude}
  \item{long}{Longitude}
  ...
}
}
\source{
\url{https://mohp.gov.np}
}
\usage{
district_val
}
\description{
A dataset with value of shape files with values of district
}
\keyword{datasets}
